Check Files in a Sequence

Check Files in a Sequence

util_check_file_seq(
  dir = getwd(),
  start = 1L,
  end = 10L,
  digits = 5L,
  fn = "filename",
  ext = "csv",
  sep = "_",
  prefix = TRUE
)

Arguments

dir
Character string. Directory.

start
Integer. Start of the sequence.

end
Integer. End of the sequence.

digits
Integer. Digits used in number sequence.

fn
Character string. Filename EXCLUDING the sequence number, the separator (sep), and the file extension (ext). The sequence number, separator and file extension are declared in the other arguments.

ext
Character string. File extension, for example ext = "csv" for Comma Separated Values.

sep
Character string. String that separates number sequence and fn.

prefix
Logical. If TRUE, number sequence are at the beginning of fn. If FALSE, number sequence are at the end of fn.

Author

Ivan Jacob Agaloos Pesigan

Examples

if (FALSE) {
# This will search for
# {5 digits}_filename.csv
util_check_file_seq(
  dir = getwd(),
  start = 1L,
  end = 10L,
  digits = 5L,
  fn = "filename",
  ext = "csv",
  sep = "_",
  prefix = TRUE
)
# This will search for
# filename_{5 digits}.csv
util_check_file_seq(
  dir = getwd(),
  start = 1L,
  end = 10L,
  digits = 5L,
  fn = "filename",
  ext = "csv",
  sep = "_",
  prefix = FALSE
)
}